# api-gateway/src/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ZyvoPay API Gateway",
    "description": "ZyvoPay Payment Gateway API - Complete banking-as-a-service platform for PIX payments, TED transfers, and financial operations.\n\n## Authentication\nAll endpoints (except `/health` and `/api/auth/clients`) require API Key authentication.\n\nProvide your API key via:\n- `X-API-Key` header, or\n- `Authorization: Bearer <api_key>` header\n\n## Rate Limiting\n- Default: 60 requests per minute per API key\n- Rate limit headers are returned in responses\n\n## Webhooks\nConfigure your webhook URL in the client settings to receive real-time event notifications.\n",
    "version": "1.0.0",
    "contact": {
      "name": "ZyvoPay Support",
      "email": "support@zyvopay.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.zyvopay.com",
      "description": "Production server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/scheduled": {
      "post": {
        "summary": "Create scheduled transfer",
        "tags": [
          "Scheduled"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "subaccount_id",
                  "transfer_type",
                  "amount_cents",
                  "scheduled_date"
                ],
                "properties": {
                  "subaccount_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "transfer_type": {
                    "type": "string"
                  },
                  "pix_key": {
                    "type": "string"
                  },
                  "pix_key_type": {
                    "type": "string"
                  },
                  "ted_bank_code": {
                    "type": "string"
                  },
                  "ted_agency": {
                    "type": "string"
                  },
                  "ted_account": {
                    "type": "string"
                  },
                  "amount_cents": {
                    "type": "integer"
                  },
                  "scheduled_date": {
                    "type": "string",
                    "format": "date"
                  },
                  "scheduled_time": {
                    "type": "string",
                    "format": "time"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Scheduled transfer created"
          }
        }
      },
      "get": {
        "summary": "List scheduled transfers",
        "tags": [
          "Scheduled"
        ],
        "parameters": [
          {
            "name": "subaccount_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of scheduled transfers"
          }
        }
      }
    },
    "/api/scheduled/{id}": {
      "delete": {
        "summary": "Cancel scheduled transfer",
        "tags": [
          "Scheduled"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scheduled transfer cancelled"
          }
        }
      }
    },
    "/api/recurring": {
      "post": {
        "summary": "Create recurring payment",
        "tags": [
          "Recurring"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "subaccount_id",
                  "pix_key",
                  "pix_key_type",
                  "frequency",
                  "amount_cents",
                  "start_date"
                ],
                "properties": {
                  "subaccount_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "pix_key": {
                    "type": "string"
                  },
                  "pix_key_type": {
                    "type": "string"
                  },
                  "frequency": {
                    "type": "string",
                    "enum": [
                      "daily",
                      "weekly",
                      "monthly",
                      "yearly"
                    ]
                  },
                  "amount_cents": {
                    "type": "integer"
                  },
                  "start_date": {
                    "type": "string",
                    "format": "date"
                  },
                  "end_date": {
                    "type": "string",
                    "format": "date"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Recurring payment created"
          }
        }
      },
      "get": {
        "summary": "List recurring payments",
        "tags": [
          "Recurring"
        ],
        "parameters": [
          {
            "name": "subaccount_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "is_active",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of recurring payments"
          }
        }
      }
    },
    "/api/recurring/{id}": {
      "patch": {
        "summary": "Update recurring payment",
        "tags": [
          "Recurring"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "is_active": {
                    "type": "boolean"
                  },
                  "end_date": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Recurring payment updated"
          }
        }
      },
      "delete": {
        "summary": "Cancel recurring payment",
        "tags": [
          "Recurring"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recurring payment cancelled"
          }
        }
      }
    },
...
  }
}